Release of version 2021.02.0

+++ gridappsd/docker_handler.py
@@ -4,8 +4,10 @@ from copy import deepcopy
 from datetime import datetime, timedelta
 import logging
 import os
+from pprint import pprint
 from subprocess import PIPE
 import threading
+from typing import Optional, Union
 
 import pkg_resources
 from pathlib import Path
@@ -76,10 +78,7 @@ if HAS_DOCKER:
             "image": "gridappsd/influxdb:{{DEFAULT_GRIDAPPSD_TAG}}",
             "pull": True,
             "ports": {"8086/tcp": 8086},
-            "environment": {"INFLUXDB_DB": "proven"},
-            "links": "",
-            "volumes": "",
-            "entrypoint": "",
+            "environment": {"INFLUXDB_DB": "proven"}
         },
         "redis": {
             "start": True,
@@ -87,8 +86,6 @@ if HAS_DOCKER:
             "pull": True,
             "ports": {"6379/tcp": 6379},
             "environment": [],
-            "links": "",
-            "volumes": "",
             "entrypoint": "redis-server --appendonly yes",
         },
         "blazegraph": {
@@ -96,10 +93,7 @@ if HAS_DOCKER:
             "image": "gridappsd/blazegraph:{{DEFAULT_GRIDAPPSD_TAG}}",
             "pull": True,
             "ports": {"8080/tcp": 8889},
-            "environment": [],
-            "links": "",
-            "volumes": "",
-            "entrypoint": "",
+            "environment": []
         },
         "mysql": {
             "start": True,
@@ -110,12 +104,10 @@ if HAS_DOCKER:
                 "MYSQL_RANDOM_ROOT_PASSWORD": "yes",
                 "MYSQL_PORT": "3306"
             },
-            "links": "",
             "volumes": {
                 data_dir + "/dumps/gridappsd_mysql_dump.sql": {"bind": "/docker-entrypoint-initdb.d/schema.sql",
                                                                "mode": "ro"}
             },
-            "entrypoint": "",
             "onsetupfn": mysql_setup
         },
         "proven": {
@@ -133,9 +125,7 @@ if HAS_DOCKER:
                 "PROVEN_IDB_USERNAME": "root",
                 "PROVEN_IDB_PASSWORD": "root",
                 "PROVEN_T3DIR": "/proven"},
-            "links": {"influxdb": "influxdb"},
-            "volumes": "",
-            "entrypoint": "",
+            "links": {"influxdb": "influxdb"}
         }
     }
 
@@ -150,15 +140,17 @@ if HAS_DOCKER:
                 "DEBUG": 1,
                 "START": 1
             },
-            "links": {"mysql": "mysql", "influxdb": "influxdb", "blazegraph": "blazegraph", "proven": "proven",
+            "links": {"mysql": "mysql",
+                      "influxdb": "influxdb",
+                      "blazegraph": "blazegraph",
+                      "proven": "proven",
                       "redis": "redis"},
             "volumes": {
                 str(Path(GRIDAPPSD_CONF_DIR).joinpath("entrypoint.sh")):
                     {"bind": "/gridappsd/entrypoint.sh", "mode": "rw"},
                 str(Path(GRIDAPPSD_CONF_DIR).joinpath("run-gridappsd.sh")):
                     {"bind": "/gridappsd/run-gridappsd.sh", "mode": "rw"}
-            },
-            "entrypoint": "",
+            }
         }
     }
 
@@ -188,6 +180,7 @@ if HAS_DOCKER:
         DEFAULT_DOCKER_DEPENDENCY_CONFIG.update(__update_template_data__(__TPL_DEPENDENCY_CONFIG__, __replace_dict__))
         DEFAULT_GRIDAPPSD_DOCKER_CONFIG.update(__update_template_data__(__TPL_GRIDAPPSD_CONFIG__, __replace_dict__))
 
+
     class Containers:
         """
         This class allows the creation/management of containers created by the gridappsd
@@ -197,11 +190,53 @@ if HAS_DOCKER:
             self._container_def = container_def
 
         @staticmethod
-        def reset_all_containers():
+        def container_list(ignore_list: Optional[Union[str, list]] = "gridappsd_dev"):
+            """
+            Provides a wrapper around the listing of docker containers.  This function was
+            brought about when running from within a docker container.
+
+            Currently the docker container that is run using docker-compose up from the
+            gridappsd-dev-environment is specified as gridappsd_dev, however this can change
+            and could potentially be extended to multiple named containers.
+
+            @param: ignore_list:
+                optional container names to not stop :: A string or list of strings
+            """
             client = docker.from_env()
 
+            if ignore_list is None:
+                ignore_list = []
+            elif isinstance(ignore_list, str):
+                ignore_list = [ignore_list]
+            containers = []
             for container in client.containers.list():
-                container.kill()
+                if container.name not in ignore_list:
+                    containers.append(container)
+            return containers
+
+        @staticmethod
+        def reset_all_containers(ignore_list: Optional[Union[str, list]] = "gridappsd_dev"):
+            """
+            Provides a wrapper around the resetting of all docker containers.  This function was
+            brought about when running from within a docker container.
+
+            Currently the docker container that is run using docker-compose up from the
+            gridappsd-dev-environment is specified as gridappsd_dev, however this can change
+            and could potentially be extended to multiple named containers.
+
+            @param: ignore_list:
+                optional container names to not stop :: A string or list of strings
+            """
+            client = docker.from_env()
+
+            if ignore_list is None:
+                ignore_list = []
+            elif isinstance(ignore_list, str):
+                ignore_list = [ignore_list]
+
+            for container in client.containers.list():
+                if container.name not in ignore_list:
+                    container.kill()
 
         def check_required_running(self, config):
             my_config = deepcopy(config)
@@ -213,7 +248,9 @@ if HAS_DOCKER:
             assert not my_config, f"The required containers were not satisfied missing {list(my_config.keys())}"
 
         def start(self):
+            pprint(DEFAULT_GRIDAPPSD_DOCKER_CONFIG)
             client = docker.from_env()
+            print(f"Docker client version: {client.version()}")
             for service, value in self._container_def.items():
                 if self._container_def[service]['pull']:
                     _log.debug(f"Pulling {service} : {self._container_def[service]['image']}")
@@ -232,17 +269,18 @@ if HAS_DOCKER:
                     kwargs['remove'] = True
                     kwargs['name'] = service
                     kwargs['detach'] = True
-                    if self._container_def[service]['environment']:
+                    if self._container_def[service].get('environment'):
                         kwargs['environment'] = value['environment']
-                    if self._container_def[service]['ports']:
+                    if self._container_def[service].get('ports'):
                         kwargs['ports'] = value['ports']
-                    if self._container_def[service]['volumes']:
+                    if self._container_def[service].get('volumes'):
                         kwargs['volumes'] = value['volumes']
-                    if self._container_def[service]['entrypoint']:
+                    if self._container_def[service].get('entrypoint'):
                         kwargs['entrypoint'] = value['entrypoint']
-                    if self._container_def[service]['links']:
+                    if self._container_def[service].get('links'):
                         kwargs['links'] = value['links']
-                    # print (kwargs)
+                    for k, v in kwargs.items():
+                        print(f"k->{k}, v->{v}")
                     container = client.containers.run(**kwargs)
                     self._container_def[service]['containerid'] = container.id
             print([x.name for x in client.containers.list()])
